+++ lizard_progress/email_notifications/signals.py
@@ -1,0 +1,12 @@
+from django.dispatch import Signal
+
+notify = Signal(
+    providing_args=[
+        'notification_type',
+        'recipient',
+        'actor',
+        'action_object',
+        'target',
+        'extra',
+    ]
+)